power: smb-lib: Fix number of VCONN attempts
The vconn_attempts counter variable currently gets incremented twice per
callback, which results in only one attempt before the counter gets
reset. This is contrary to the intended logic of checking for
VCONN_MAX_ATTEMPTS before giving up.

Fix this by removing the redundant increment operation. And while we're
at it, print the right variable in the debug error message as well.

CRs-Fixed: 2196518
Change-Id: I30983ca282703766ba90c4918a8bf97767b8108d
Signed-off-by: Guru Das Srinagesh <gurus@codeaurora.org>
Signed-off-by: Udit Karode <udit.karode@gmail.com>