EXPERIMENTAL: ANDROID: cpufreq: opp: Use dynamic-power-coefficient to register an EM
The "dynamic-power-coefficient" in DT can be used to estimate the power
consumed by CPUs at any OPP. Use it to register an Energy Model when
using the cpufreq-dt and arm_big_little CPUFreq drivers.

The code that performs the actual power estimation is factored out in
PM_OPP, hence making it easy to use in other drivers if need be.

Change-Id: Ia808262ef6c9f2cc7819a83e8eb2f602454edfa3
Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Udit Karode <udit.karode@gmail.com>