f2fs: rename /sys/fs/f2fs
fsmgr and vold tries to read nodes from here and run GC by itself.

Since it conflicts with rapid GC(not to mention less effective),
rename /sys/fs/f2fs to /sys/fs/f2fs_dev to avoid Android reading it entirely.

We could rename gc_urgent and dirty_segments but this approach seems better
as it'll still be detected as a f2fs volume and try to run GC.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Panchajanya1999 <rsk52959@gmail.com>
Signed-off-by: Udit Karode <udit.karode@gmail.com>