sched/tune: Introduce SchedTune Assist
Based on YaroST12/Z2_Kernel@eeddf47:

This implements a mechanism by which default SchedTune parameters
can be configured in-kernel, circumventing userspace, and
mitigating reliance on ramdisk modification in the context of
custom kernels.

This is a V2.5 version that adds proper protection
from userspace (mainly init) trying to write lame
boost values and gives full control to developer
and user (sh is not blocked).

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
[0ctobot: Update for msm-4.9 and improve coding style]
[yaro]: Update for 4.14
Co-authored-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: Udit Karode <udit.karode@gmail.com>